Don't apply prefetching logic when prefetch_us == 0
This fixes the race condition that caused the ClusterReader to sometimes edge
ahead of the MuonReader, causing IBDs to leak through which should have been
vetoed. Now we're getting LBNL-consistent IBDs for 21221.1. However, prefetch_us
and epsilon_us should be scrapped in favor of a "lead time" and "lag time", or a
signed "delta_us". The current design is very confusing.